+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # 2026/5/28
 
 - Char-Level Tokenizer
+- MLM Test
+- MLM Backward (only update W_vocab & b_vocab) => best loss: 2.9587
 
 # 2026/5/26